경북대 BE 오성식 Step3-위시 리스트 (#334)
* refactor: 커스텀 예외 생성 및 예외 처리 위치 변경

* refactor: 반환타입 Optional로 변경

* refactor: 로그인 예외 및 validation 추가

* refactor: getter 지양 -> user의 내부 메서드로 구현

* refactor: Dao 수정 메서드 책임 분리

* refactor: 비밀번호 해싱(salt+PBKDF2)하여 저장하도록 변경

* docs: README.md 수정

* feat: 위시리스트 조회 기능

* feat: 위시리스트 추가 기능

* feat: 위시리스트 삭제 기능

* refactor: Filter 피드백 반영(매직넘버, 헤더 추출)

* refactor: 상수 적용

* refactor: requestDto userId 제외

* refactor: exception 처리 변경

* refactor: exception 상위로 던지는 부분 제거

* refactor: EmptyResultDataAccessException 변경

* refactor: assignableTypes 변경

* refactor: 생성자 변경

+++ src/main/java/gift/user/service/UserService.java
@@ -1,6 +1,7 @@
 package gift.user.service;
 
-import gift.product.domain.Product;
+import gift.auth.PasswordUtil;
+import gift.common.exception.NoSuchIdException;
 import gift.user.domain.User;
 import gift.user.dto.UserPatchRequestDto;
 import gift.user.dto.UserSaveRequestDto;
@@ -8,7 +9,9 @@ import gift.user.repository.UserDao;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Base64;
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 @Service
@@ -17,10 +20,14 @@ public class UserService {
     public UserService(UserDao userDao) {
         this.userDao = userDao;
     }
+
     @Transactional
     public User save(UserSaveRequestDto userSaveRequestDto) {
+
         UUID uuid = UUID.randomUUID();
-        User user = new User(uuid, userSaveRequestDto.getEmail(), userSaveRequestDto.getPassword());
+        byte[] salt = PasswordUtil.generateSalt();
+        String hashedPassword = PasswordUtil.encryptPassword(userSaveRequestDto.getPassword(), salt);
+        User user = new User(uuid, userSaveRequestDto.getEmail(), hashedPassword, Base64.getEncoder().encodeToString(salt));
         return userDao.save(user);
     }
 
@@ -30,24 +37,37 @@ public class UserService {
     }
 
     @Transactional(readOnly = true)
-    public User findById(UUID id) {
+    public Optional<User> findById(UUID id) {
         return userDao.findById(id);
     }
 
     @Transactional(readOnly = true)
-    public User findByEmail(String Email) {
+    public Optional<User> findByEmail(String Email) {
         return userDao.findByEmail(Email);
     }
 
     @Transactional
     public User updateUser(UUID id, UserPatchRequestDto userPatchRequestDto) {
-        userDao.findById(id);
-        return userDao.update(id, userPatchRequestDto);
+
+        if (userDao.findById(id).isEmpty()) {
+            throw new NoSuchIdException("존재하지 않는 ID 입니다.");
+        }
+        if (userPatchRequestDto.getEmail() != null) {
+            userDao.updateEmail(id, userPatchRequestDto.getEmail());
+        }
+        if (userPatchRequestDto.getPassword() != null) {
+            byte[] salt = Base64.getDecoder().decode(userDao.findById(id).get().getSalt());
+            String hashedPassword = PasswordUtil.encryptPassword(userPatchRequestDto.getPassword(), salt);
+            userDao.updatePassword(id, hashedPassword);
+        }
+        return userDao.findById(id).get();
     }
 
     @Transactional
     public void deleteUser(UUID id) {
-        userDao.findById(id);
+        if(userDao.findById(id).isEmpty()) {
+            throw new NoSuchIdException("존재하지 않는 ID입니다.");
+        }
         userDao.delete(id);
     }
 }
